Add files via upload
Statistics: A way to get information from data.
Descriptive Statistics: Deals with methods of organizing, summarizing, and presenting data in a convenient and informative way.
Measures of Central location

Mean: 𝜇=∑𝑖=1𝑁𝑥𝑖𝑁
Median: The median is calculated by placing all the observations in ascending or descending order.
N is Odd: The observation that falls in the middle is the median. (𝑛+12)𝑡ℎ item.
N is Even: Median is determined by averaging the two observations in the middle, (𝑛2)𝑡ℎ+(𝑛2+1)𝑡ℎ2 item.
Mode: Observations that occurs with the greatest frequency.

Measures of Variability/Dispersion

Variance: How far each observation is from the mean. These differences from the mean are called deviations.
The average squared deviation from the mean is called the variance.  𝜎2=∑𝑖=1𝑁(𝑥𝑖−𝜇)2𝑁 
Range: Largest obs-Smallest obs

Empirical Rule

Approximately 68% of all observations fall within one standard deviation of the mean. (mean+-1std)
Approximately 95% of all observations fall within two standard deviations of the mean. (mean+-2std)
Approximately 99.7% of all observations fall within three standard deviations of the mean. (mean+-3std)

Basics of Probability

Random Experiment: An action that leads to one of several possible outcomes.Ex: Tossing a coin, Covid Test
Event: An event is the outcome of an experience. Ex: Head
Trail: Each time a random experiment is conducted, it is called a trial.
Sample Space: A list of all possible outcomes of an experiment. Ex: {H,T} , {+ve,-ve}
Probability: How likely is that an event can occur. Let's A is an event.

P(A) = Favorable number of events/total number of events.
Let A be an event Then 0<=P(A)<=1.
Random Variable: A rule that assigns a number to each outcome of an experiment. Ex: 1 head or 1 tail, 2 heads

Discrete and Continuous Random Variable

Discrete Random Variable: Which takes countable number of values. Ex: #Customers in a queue, #Products added to cart
Continuous Random Variable: Whose values are uncountable. Ex: Amount of time to complete a task